Re-work to take startTime to separate "initial" event
The idea is that when implementing live scheduling, the library
will need to ask the consumer whether to schedule an event in
response to some stream info; however, the start time will not be
defined by the consumer (the stream will decide this), and so it
will not make sense to expose a startTime property here.

Therefore, the startTime scheduling only makes sense for manual
interstitial scheduling, which for now I intend to only support at
stream start (hence "initial"), for a VOD interstitial scheduling
experience.