Change tactic and go for simple VOD insertion first
The plan is to base insertion entirely off of date-range and
program-date-time notified in the individual playlists. For live
content, a mis-match in start-date values would indicate that the
packager had placed the date-range separately for the different
playlists, which would seem wrong to me at least. Therefore, seems
this is the easiest truth to adhere to.

For VOD, given that we want to define interstitials in terms of
absolute time-values, in order to avoid mis-match between video
and audio, I make the following assumptions:
  1) Insertion will only occur on EXT-X-PLAYLIST-TYPE:VOD
  2) Insertion will not occur if EXT-X-PROGRAM-DATE-TIME is already
     present
  3) All playlists refer to the same starting point

These assumptions limit the scope of where the ad insertion will
happen, but hopefully it still covers the majority of cases, and
this can be improved in future.