Changed public API so pre-roll and initial mid-roll are provided in delegate.
This change is made because it was difficult to have context of
what kind of asset (VOD vs live) was being loaded when providing
pre-roll interstitials (which is important as the resume offset
should be indefinite for live and zero for VOD).

As part of this change I also introduced a new (hopefully more
robust) way of handling event decisioning. This new method makes
use of Swift Concurrency and so the minimum target of the package
has been bumped to v13. Swift Concurrency was chosen as there is a
decent amount of asynchronous waiting going on while making
insertions into the manifest in response to AVFoundation loading
requests, and Swift Concurrency makes handling these journeys much
more convenient.