Implement X-ASSET-LIST handling for iOS 16.1+
As of iOS 16.1 we are able to intercept X-ASSET-LIST requests via
AVAssetResourceLoaderDelegate. This allows us to use this approach
instead of multiple tags each with their own X-ASSET-URI. Their may
not be much benefit to this approach; however, it is more in line
with how a backend server may deliver an interstitial response, and
so we prefer this usage over the former to better approximate how
a server driven approach may look like.

In addition, some bugs with the event decisioning actor were fixed.